add anodized-macros crate
