+++ crates/anodized-macros/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "anodized"
-description.workspace = true
+name = "anodized-macros"
+description = "Proc macros used by the Anodized specification system"
 
 version.workspace = true
 edition.workspace = true
@@ -10,12 +10,14 @@ categories.workspace = true
 keywords.workspace = true
 license.workspace = true
 
+[lib]
+proc-macro = true
+
 [lints.rust]
 unexpected_cfgs = { level = "warn", check-cfg = ["cfg(anodized_panic)", "cfg(anodized_print)"] }
 
 [dependencies]
-anodized-macros.workspace = true
-
-[dev-dependencies]
-pollster = "0.4"
-trybuild = "1.0"
+anodized-core.workspace = true
+proc-macro2.workspace = true
+quote.workspace = true
+syn.workspace = true
