+++ crates/anodized-macros/README.md
@@ -1,0 +1,5 @@
+<img width="100" alt="Anodized Logo" src="https://raw.githubusercontent.com/mkovaxx/anodized/main/assets/logo.svg">
+
+# Anodized Macros
+
+You're probably looking for the [`anodized`](https://crates.io/crates/anodized) crate, which re-exports the proc macros provided by this crate, to work around the limitation that [a `proc-macro` crate can only export proc macros](https://doc.rust-lang.org/reference/linkage.html#r-link.proc-macro).
